implement repository ipmlementation for team,player and tournoi

+++ src/main/java/org/GestionDesTournois/Repository/Interfaces/TournoiInterface.java
@@ -11,5 +11,5 @@ public interface TournoiInterface {
     boolean deleteTournoi(int id);
     List<Tournoi> getAllTournois();
     Optional<Tournoi> getTournoiById(int id);
-    int calculerDureeEstimeeTournoi(int tournoiId);
+    double calculerDureeEstimeeTournoi(int tournoiId);
 }
